requirements.txt
streamlit==1.22.0
opencv-python==4.7.0.72
numpy==1.24.2
matplotlib==3.7.1
Pillow==9.5.0